requirements.txt
requests>=2.32.0
openai>=2.36.0
python-dotenv>=1.2.2
Django>=5.2.0